refactor: move additional response methods
